+++ R/checks.R
@@ -77,38 +77,3 @@
     return( query )
   }
 }
-
-.check_drive <- function(registry){
-  if (.Platform$OS.type != "windows"){
-    warning("Auxiliary function .check_drive has been called. Not necessary nor ",
-            "appropriate if not on Windows.")
-    return(NA)
-  }
-  registry_drive <- path_get_drive_letter(registry)
-  wd_drive <- path_get_drive_letter(getwd())
-  if (registry_drive != wd_drive){
-    default_working_dir <- getwd()
-    if (interactive()){
-      msg <- c("The drive of your current working directory differs from the drive of",
-               "the registry directory. The corpus data cannot be processed by CWB tools",
-               "unless the drive is set to", registry_drive, "- please confirm if you",
-               "agree to resetting the working directory [Y/n]:"
-      )
-      user_feedback <- readline(prompt = paste0(msg, collapse = " "))
-      if (user_feedback == "Y"){
-        default_working_dir <- getwd()
-        setwd(registry_drive)
-        return(default_working_dir)
-      } else {
-        warning("Drive of working directory differs from drive of registry directory. ",
-                "CWB may fail to find data directory.")
-        return(NA)
-      }
-    } else {
-      warning("Drive of working directory differs from drive of registry directory. ",
-              "CWB may fail to find data directory.")
-    }
-  } else {
-    return(NA)
-  }
-}